+++ manual/source/version_history.rst
@@ -173,6 +173,8 @@ Bug Fixes
 * Fix uninitialised variable in BDSBunch.
 * Fix float / double casts in sampler output.
 * Fix possible bad access by indexing beyond range of array in dipole fringe integrator.
+* The maximum step length in a muon spoiler is now 1/20th of the length whereas before it was the full length.
+  This step limit applies only in the 'yoke' (i.e. the outer part) of the spoiler and not in the pipe part.
 
 
 Output Changes